Ptrlabs (ptrlabs v0.1.0)

Provides methods for passing the PTR course.

Functions

arrangements(n, r)

comb(list, k, fList)

cylinderArea(h, r)

Returns the area of a cylinder given the radius and the height

Examples

  iex> Ptrlabs.cylinderArea(3,4)
  175.92904

decode(ciphertext, key)

Returns the plaintext of a caesar cipher encoding given a ciphertext and a key.Works with upper and lower case.

Examples

  iex>Ptrlabs.decode("oRuHp",3)
  "lOrEm"
  iex>Ptrlabs.decode("oruhp",3)
  "lorem"

encode(plaintext, key)

Returns the ciphertext of a caesar cipher encoding given a plaintext and a key.Works with upper and lower case.

Examples

  iex>Ptrlabs.encode("lorem",3)
  "oruhp"
  iex>Ptrlabs.encode("lOrEm",3)
  "oRuHp"

extractRandomN(list, n)

Returns a list of n elements randomly selected from a given list

Examples

  iex>test = Ptrlabs.extractRandomN([1,2,3,4,8,4],3)
  iex>length(test) == 3 && MapSet.subset?(MapSet.new(test),MapSet.new([1,2,3,4,8,4]))
  true

factorial(number)

factorize(number)

Returns the prime factors of a given number

Examples

  iex>Ptrlabs.factorize(42)
  [2,3,7]

factors(n)

fib(list, n)

firstFibonacciElements(n)

Return n fibonacci numbers,starting from 1

Examples

  iex>Ptrlabs.firstFibonacciElements(7)
  [1,1,2,3,5,8,13]

getCase(char)

greetPTR()

Returns an honest greeting.

Examples

  iex> Ptrlabs.greetPTR
  "hello PTR"

letterCombinations(text)

lineWords(list)

Given a list of strings returns the ones that could be written using only the buttons from one row of a qwerty keyboard

Examples

  iex>Ptrlabs.lineWords(["hello","alaska","dad","peace"])
  ["alaska","dad"]

listRightAngleTriangles()

Returns a list of all the possible values for the lengths of the sides of a valid right angled triangle, where the length of the base and the perpendicular can't exceed 20.

prime?(n)

Checks if a given number is prime

Examples

  iex> Ptrlabs.prime?(3)
  true

removeConsecutiveDuplicates(list)

Returns a given list without consecutive duplicates

Examples

  iex>Ptrlabs.removeConsecutiveDuplicates([1,2,2,2,4,8,4])
  [1,2,4,8,4]

reverse(list)

Returns the reverse of a given list

Examples

  iex> Ptrlabs.reverse([1,2,3])
  [3,2,1]

rotateLeft(list, count)

Rotates a given list count number of places and returns it.

Examples

  iex>Ptrlabs.rotateLeft([1,2,4,8,4],3)
  [8,4,1,2,4]
  iex>Ptrlabs.rotateLeft([1,2,4,8,4],5)
  [1,2,4,8,4]
  iex>Ptrlabs.rotateLeft([1,2,4,8,4],101)
  [2,4,8,4,1]

shift(minmax, char, key)

smallestNumber(list)

Given a list of numbers returns the smallest number that can be obtained by arranging the numbers. A number can't start with a 0

Examples

  iex>Ptrlabs.smallestNumber([3,2,1])
  123
  iex>Ptrlabs.smallestNumber([0,5,1])
  105
  iex>Ptrlabs.smallestNumber([])
  "Can't be empty"

translator(dictionary, original_string)

Given a dictionary and a string returns a string containing the translated version of all the words that matched. A word is considered a substring that is followed by a ' ' character. If the first argument is not a map the default dictionary will be used stored in the module as @dictionary.

Examples

  iex> Ptrlabs.translator(1,"mama is with papa")
  "mother is with father"

uniqueSum(list)

Returns the sum of unique elements of a given list

Examples

  iex> Ptrlabs.uniqueSum([1,2,4,8,4,2])
  15